refactor(utils): give the last four symbols honest homes
The previous commit parked str2bool / fmt / STD_SCALE / download_from_huggingface
in utils/utils.py. That was wrong: utils.py is the hydra + lightning experiment
lifecycle module (extras, task_wrapper, get_metric_value), so this just traded one
grab-bag for another.

  str2bool                  -> type_utils.py  (a str->bool coercion, next to the
                               existing _to_numpy coercion)
  download_from_huggingface -> hf_utils.py    (new; keeps the Hub dependency
                               visible and optional, mirroring the aws/ subpackage)
  STD_SCALE                 -> algo/hpt.py    (not a shared utility at all -- it is
                               the HPT action-token init scale with exactly one
                               consumer, 20 lines below where it now lives)
  fmt                       -> deleted

fmt turned out to be dead. It was classified as live off a reference count that
word-matched unrelated  keyword arguments in plot_cartesian.py and in
vendored lerobot. Nothing has ever imported it -- not on main, not now.

utils.py is back to its three lifecycle helpers.

Verified: ruff F401/F821/F811 clean on egomimic/utils and algo/hpt.py; 13/14
affected modules import (the 14th is robot/rollout, pending #563).
