+++ egomimic/algo/hpt.py
@@ -17,7 +17,12 @@ from egomimic.algo.algo import Algo
 from egomimic.models.hpt_nets import MultiheadAttention, SimpleTransformer
 from egomimic.rldb.embodiment.embodiment import get_embodiment, get_embodiment_id
 from egomimic.utils.tensor_utils import EinOpsRearrange, get_sinusoid_encoding_table
-from egomimic.utils.utils import STD_SCALE, download_from_huggingface
+from egomimic.utils.hf_utils import download_from_huggingface
+
+
+# Init scale for the HPT action tokens (was a shared constant in utils.py;
+# it has exactly one consumer, below).
+STD_SCALE = 0.02
 
 
 class HPTModel(nn.Module):
