+++ egomimic/utils/type_utils.py
@@ -1,3 +1,4 @@
+import argparse
 import numpy as np
 
 
@@ -9,3 +10,14 @@ def _to_numpy(arr):
     if hasattr(arr, "numpy"):
         return arr.numpy()
     return np.asarray(arr)
+
+
+def str2bool(value):
+    if isinstance(value, bool):
+        return value
+    value = value.lower()
+    if value in ("yes", "true", "t", "y", "1"):
+        return True
+    if value in ("no", "false", "f", "n", "0"):
+        return False
+    raise argparse.ArgumentTypeError("Boolean value expected.")
